Support mounting D41 images with up to 42 tracks
To be recognized as a valid D41 image, the file size has to be
between 174848 (size for a standard 35 track image without
error info) and 206114 bytes (size for a 42 track image with 17
sectors for tracks >35 plus 1 byte error info for each of these
802 sectors).

The main motivation for this change is to make it possible to
mount such images and access the data on tracks <=35. While it's
possible to read and write sectors beyond track 35 with B-R/B-W
commands, they are not considered when allocating new sectors.
Reading files which extend there only works if supported by the
respective loader.