Enable custom, temporary magic values for M-R handling
This enables loader handlers to set custom magic values in
response to specific events at runtime.

If the address in a M-R command matches the value set in the
global "custom_magic" variable, it uses the values from there
instead of a potential entry in c1541_magics[].

The address entry in "custom_magic" is cleared after the next
M-R command, even if it did *not* reference the magic's address.
This might not work in cases when the result is fetched over
multiple reads. But for now be safe and maybe re-evaluate if
such cases are found.