Enable hidden files in D64 images
Treats directory entries with a zero file type but track != 0 as
"hidden". These entries are then included in the enumeration by
next_match() if called with FLAG_HIDDEN set for the filetype.

Has to be enabled by setting the d64dh.hidden flag *after* calling
opendir().

Warning: As d64dh is the member of a *union* in dh_t, this is only
valid after confirming that a D64 image is actually mounted.
Otherwise this will garble the active directory handle!