Revert "Default to 0 for first two M-R result bytes"
This reverts commit 26dd93eb3a69ba03f808253b5984603389eba525.

Turns out that:

1. The content in error_buffer[] isn't really random during M-R
   command handling, but always "00, OK,00,00", because the error
   buffer is reset just before parsing the command.
2. "GEOS MegaPatch 3" relies on this behavior for sd2iec detection.

As after all the result for M-R to unhandled memory locations
already seems to be predictable with the original implementation,
revert the commit in the interest of the least intrusive change
in respect to the official firmware.

This also requires minor changes to the "Krill's loader" handler,
as it depends on the returned value during the loader's drive
probing.